backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy>=2.0
aiosqlite
asyncpg
pyjwt
pydantic-settings
pwdlib[argon2]